删除重复文件控制
