+++ core/config.ini
@@ -21,6 +21,8 @@ ignore_video_file_less_than = 232
 enable_file_move = yes
 # 传入正确的ffprobe 路径会简单检查文件是否合法媒体文件
 document_completion =
+# 是否删除重复文件
+is_delect_duplicate_file=
 
 [Network]
 # 是否启用代理
